+++ src/main/java/FrontEndUtilities/GUIDataMaster.java
@@ -2,9 +2,11 @@ package FrontEndUtilities;
 
 import BackEndUtilities.Sample;
 import GUI.Card;
-import GUI.CardTypes;
+import Enums.CardTypes;
 import GUI.Frame;
-import Graphing.GraphManager;
+import Managers.ErrorManager;
+import Managers.GraphManager;
+import Managers.OutputManager;
 import TableUtilities.DataTable;
 
 import java.util.ArrayList;
