+++ src/main/java/GUI/Panels/MiddlePanelFive.java
@@ -1,23 +1,28 @@
-package GUI;
+package GUI.Panels;
 
-import FrontEndUtilities.ErrorManager;
+import Enums.CardTypes;
+import GUI.Card;
+import GUI.CellsTable;
+import Managers.ErrorManager;
 import FrontEndUtilities.GUIDataMaster;
 
 import javax.swing.*;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class MiddlePanelTwo extends Card{
-    private final CardTypes type = CardTypes.TWO_DATA_NO_VARIABLE;
-
-    private static JTextArea dataArea0;
-    private static JTextArea dataArea1;
+public class MiddlePanelFive extends Card {
+    private static final CardTypes type = CardTypes.ONE_DATA_ONE_VARIABLE;
+    private JTextArea dataArea0;
+    private JTextField variableArea0;
 
     private JLabel dataLabel0;
-    private JLabel dataLabel1;
+    private JLabel variableLabel0;
 
-    public MiddlePanelTwo(){
+    public MiddlePanelFive(){
+        /*Create a JPanel with a grid bag layout*/
         this.setLayout(new GridBagLayout());
 
         /*Create the constraints for gridbag layout and apply them to the scroll pane.*/
@@ -26,11 +31,11 @@ public class MiddlePanelTwo extends Card{
         c.anchor = GridBagConstraints.LINE_END;
         c.gridx = 2;
         c.gridy = 1;
-        this.add(middlePanel(), c);
+        this.add(scrollPane(), c);
         c.gridy = 2;
         this.add(topPanel2(), c);
         c.gridy = 3;
-        this.add(middlePanel2(), c);
+        this.add(variableArea(), c);
 
         /*Change the y position value for the gridbag constraints and apply to the panel containing the
          * button and label.*/
@@ -40,33 +45,61 @@ public class MiddlePanelTwo extends Card{
 
     /**Creates a scroll pane for the text area.
      *@return A scroll pane containing a text area.*/
-    private JScrollPane middlePanel() {
+    private JScrollPane scrollPane() {
         JScrollPane pane = new JScrollPane(dataArea());
-        pane.setHorizontalScrollBarPolicy(ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
-        return (pane);
-    }
-
-    private JScrollPane middlePanel2() {
-        JScrollPane pane = new JScrollPane(dataArea2());
         pane.setHorizontalScrollBarPolicy(ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
         return (pane);
     }
     /**Creates a text area that user can input data into.
      *@return The text area*/
     private JTextArea dataArea(){
-        dataArea0 = new JTextArea("Select Data from Chart", 10, 20);
+        dataArea0 = new JTextArea("Select Data from Chart", 20, 15);
         dataArea0.setEditable(false);
         dataArea0.setLineWrap(true);
+
+        dataArea0.getDocument().addDocumentListener(new DocumentListener() {
+            @Override
+            public void insertUpdate(DocumentEvent e) {
+                updateMeasureData();
+            }
+
+            @Override
+            public void removeUpdate(DocumentEvent e) {
+                updateMeasureData();
+            }
+
+            @Override
+            public void changedUpdate(DocumentEvent e) {
+
+            }
+        });
 
         return(dataArea0);
     }
 
-    private JTextArea dataArea2(){
-        dataArea1 = new JTextArea("Select Data from Chart", 10, 20);
-        dataArea1.setEditable(false);
-        dataArea1.setLineWrap(true);
+    private JTextField variableArea(){
+        variableArea0 = new JTextField();
+        variableArea0.setEditable(false);
+        variableArea0.setBackground(Color.WHITE);
 
-        return(dataArea1);
+        variableArea0.getDocument().addDocumentListener(new DocumentListener() {
+            @Override
+            public void insertUpdate(DocumentEvent e) {
+                updateMeasureData();
+            }
+
+            @Override
+            public void removeUpdate(DocumentEvent e) {
+                updateMeasureData();
+            }
+
+            @Override
+            public void changedUpdate(DocumentEvent e) {
+
+            }
+        });
+
+        return(variableArea0);
     }
 
     /**Method which creates panel to contain the button and label.
@@ -90,40 +123,15 @@ public class MiddlePanelTwo extends Card{
     /**Method which creates the label.
      *@return The label*/
     private JLabel dataLabel(){
-        return(new JLabel("Data"));
+        return(dataLabel0 = new JLabel("Data"));
     }
 
     private JLabel dataLabel2(){
-        return(new JLabel("Data"));
+        return(variableLabel0 = new JLabel("Data"));
     }
 
     /**Method which creates the button for importing data from chart.
      *@return The import button,*/
-    private JButton importButton2(){
-        JButton btnImport = new JButton("Import From Chart");
-
-        btnImport.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String d = CellsTable.getSelectedData();
-
-                if(d == null){
-                    return;
-                }
-
-                //If Shift is held when clicking
-                if((e.getModifiers() & 1) != 0){
-                    dataArea1.append("," + d);
-                }else{
-                    dataArea1.setText(d);
-                }
-                updateMeasureData();
-            }
-        });
-
-        return(btnImport);
-    }
-
     private JButton importButton(){
         JButton btnImport = new JButton("Import From Chart");
 
@@ -142,8 +150,21 @@ public class MiddlePanelTwo extends Card{
                 }else{
                     dataArea0.setText(d);
                 }
+                updateMeasureData();
+            }
+        });
 
-                updateMeasureData2();
+        return(btnImport);
+    }
+
+    private JButton importButton2(){
+        JButton btnImport = new JButton("Import From Chart");
+
+        btnImport.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                variableArea0.setText(CellsTable.getSelectedData());
+                updateMeasureDataTwo();
             }
         });
 
@@ -151,17 +172,6 @@ public class MiddlePanelTwo extends Card{
     }
 
     private void updateMeasureData(){
-        String[] data = dataArea1.getText().split(",");
-
-        int s = RightPanel.getCurrentMeasureIndex();
-        if(s < 0){
-            return;
-        }
-
-        GUIDataMaster.getGUIMeasure(s).addData(false, 1, data);
-    }
-
-    private void updateMeasureData2(){
         String[] data = dataArea0.getText().split(",");
 
         int s = RightPanel.getCurrentMeasureIndex();
@@ -170,6 +180,17 @@ public class MiddlePanelTwo extends Card{
         }
 
         GUIDataMaster.getGUIMeasure(s).addData(false, 0, data);
+    }
+
+    private void updateMeasureDataTwo(){
+        String[] data = variableArea0.getText().split(",");
+
+        int s = RightPanel.getCurrentMeasureIndex();
+        if(s < 0){
+            return;
+        }
+
+        GUIDataMaster.getGUIMeasure(s).setVariable(variableLabel0.getText(), data[0]);
     }
 
     @Override
@@ -181,27 +202,31 @@ public class MiddlePanelTwo extends Card{
     public void setDataArea(int index, String data) {
         switch (index) {
             case 0 -> dataArea0.setText(data);
-            case 1 -> dataArea1.setText(data);
             default -> ErrorManager.sendErrorMessage("GUI", "Program attempted to set data for a data field which does not exist");
         }
     }
 
     @Override
     public void setVariableArea(int index, String data) {
-        ErrorManager.sendErrorMessage("GUI", "Program attempted to set data for a data field which does not exist");
+        switch (index) {
+            case 0 -> variableArea0.setText(data);
+            default -> ErrorManager.sendErrorMessage("GUI", "Program attempted to set data for a data field which does not exist");
+        }
     }
 
     @Override
     public void setDataLabel(int index, String label) {
         switch (index) {
             case 0 -> dataLabel0.setText(label);
-            case 1 -> dataLabel1.setText(label);
             default -> ErrorManager.sendErrorMessage("GUI", "Program attempted to set name for a label which does not exist");
         }
     }
 
     @Override
     public void setVariableLabel(int index, String label) {
-        ErrorManager.sendErrorMessage("GUI", "Program attempted to set name for a label which does not exist");
+        switch (index) {
+            case 0 -> variableLabel0.setText(label);
+            default -> ErrorManager.sendErrorMessage("GUI", "Program attempted to set name for a label which does not exist");
+        }
     }
 }
