+++ src/main/java/Graphing/HorizontalBarGraph.java
@@ -1,6 +1,10 @@
 package Graphing;
 
+import Enums.DataFormat;
+import Enums.GraphTypes;
 import FrontEndUtilities.GUIMeasure;
+import Interfaces.IGraph;
+import Managers.GraphManager;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.ChartPanel;
 import org.jfree.chart.JFreeChart;
@@ -13,7 +17,7 @@ import org.jfree.data.category.DefaultCategoryDataset;
 import java.util.ArrayList;
 import java.util.TreeMap;
 
-public class HorizontalBarGraph implements IGraph{
+public class HorizontalBarGraph implements IGraph {
     private final GraphTypes type = GraphTypes.HORIZONTAL_BAR;
     private String data = "";
     private DataFormat dataFormat = DataFormat.MX_PLUS_B;
