+++ src/main/java/Managers/ErrorManager.java
@@ -1,4 +1,4 @@
-package FrontEndUtilities;
+package Managers;
 
 import javax.swing.*;
 import java.awt.*;
