+++ src/main/java/WaspAnalyzer/Main.java
@@ -1,8 +1,9 @@
 package WaspAnalyzer;
 
 import BackEndUtilities.*;
-import Graphing.GraphManager;
-import Respository.RepositoryManager;
+import Managers.GraphManager;
+import Managers.MeasureManager;
+import Managers.RepositoryManager;
 
 import Settings.Themes;
 import Settings.UserSettings;
@@ -10,6 +11,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.Arrays;
 
 public class Main {
     private static final Logger logger = LogManager.getLogger(Main.class);
@@ -17,15 +19,23 @@ public class Main {
     public static void main(String[] args) throws InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         logger.debug("Starting ApplicationMain.Main.");
 
-        Themes.init();
-        UserSettings.init();
-        RepositoryManager.init();
-        MeasureManager.init();
-        DynamicJavaClassLoader.init();
-        GraphManager.init();
-
-        GUI.Frame frame = new GUI.Frame();
-        FrontEndUtilities.GUIDataMaster.setFrameReference(frame);
+        if(Arrays.asList(args).contains("--template") || Arrays.asList(args).contains("-t")) {
+            UserSettings.init();
+            RepositoryManager.init();
+            MeasureManager.init();
+            DynamicJavaClassLoader.init();
+            return;
+        }
+        else {
+            Themes.init();
+            UserSettings.init();
+            RepositoryManager.init();
+            MeasureManager.init();
+            DynamicJavaClassLoader.init();
+            GraphManager.init();
+            GUI.Frame frame = new GUI.Frame();
+            FrontEndUtilities.GUIDataMaster.setFrameReference(frame);
+        }
 
         logger.debug("Leaving ApplicationMain.Main.");
     }
